requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
duckdb>=0.10.0
pytest>=8.0.0
pyarrow>=15.0.0
streamlit>=1.33.0
plotly>=5.20.0
